requirements.txt
websockets>=10.4
python-telegram-bot>=20.0
slackclient>=2.9.4
aiohttp>=3.8.4
pandas>=1.5.3
numpy>=1.23.5
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
mypy>=1.4.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
